Add files via upload
This project is focused on Sentiment Analysis of Tweets, analyzing the emotions and opinions expressed in tweets. The objective is to classify tweets into categories such as positive, negative, or neutral based on their sentiment. This project can be applied to understanding public opinion on topics, brands, or events using real-time Twitter data.